refactor: allow single-file benchmarks (#3)
* refactor: allow single-file benchmarks

* docs: document test architecture

---------


test: tweak a filename

build: publish ignores tests

build: publish ignores docs

build: stop ignoring on publish

test: add a todo test

docs: minor tweaks

docs: minor tweaks

build: version bump

+++ bin/builder/index.js
@@ -7,50 +7,52 @@ const replaceTokensInFile = async ({
   srcfolder, 
   filepath, 
   entrypath,
-  fragments 
+  tokens 
 }) => {    
-  for (const fragment of fragments) {
-    const srcpath = path.join(dirname, `../${srcfolder}/${fragment.target}`)
+  for (const token of tokens) {
+    const srcpath = path.join(dirname, `../${srcfolder}/${token.target}`)
     const contents = await fs.readFile(srcpath, 'utf8')
-    const processed = contents
-      .replaceAll('{{entrypath}}', entrypath)
+    const processed = contents.replaceAll('{{entrypath}}', entrypath)
 
     const existing = await fs.readFile(filepath, 'utf8')
-
-    const n1 = existing.indexOf(fragment.startToken)
-    const n2 = existing.indexOf(fragment.endToken, n1)
+    const n1 = existing.indexOf(token.start)
+    const n2 = existing.indexOf(token.end, n1)
     const rm = existing.slice(0, n1) + existing.slice(n2)
+    const res = `${rm.slice(0, n1)}${token.start}\n${processed}${rm.slice(n1)}`
 
-    await fs.writeFile(
-      filepath, 
-      `${rm.slice(0, n1)}${fragment.startToken}\n${processed}${rm.slice(n1)}`
-    )
+    await fs.writeFile(filepath, res)
+    
+    console.log('replaced token:', `"${token.start}"`, 'in:', filepath)
   }
 }
 
 const createExample = async ({ 
   srcfolder, 
-  targetfolder, 
+  targetfolder = '', 
   entrypath, 
   fragments 
 }) => {
   try {
     await fs.rm(targetfolder, { recursive: true, force: true })
-    await fs.mkdir(targetfolder)
+    if (targetfolder) await fs.mkdir(targetfolder)
     
-    console.log('created:', targetfolder)
-
     for (const fragment of fragments) {
       const srcpath = path.join(dirname, `../${srcfolder}/${fragment.target}`)
+      const targetpath = targetfolder 
+        ? `${targetfolder}/${fragment.target}` 
+        : fragment.target
       const contents = await fs.readFile(srcpath, 'utf8')
       const processed = contents.replaceAll('{{entrypath}}', entrypath).trim()
 
-      await fs.writeFile(`${targetfolder}/${fragment.target}`, processed)
+      await fs.writeFile(targetpath, processed)
+      
+      console.log('created:', targetpath)
     }    
   } catch (err) {
     console.log('An error occured.', 'cleaning up ...')
     
-    await fs.rm(targetfolder, { recursive: true, force: true })
+    if (targetfolder)
+      await fs.rm(targetfolder, { recursive: true, force: true })
     
     throw err 
   }
